Added study Blueprint support.
Allows for setting paths, data labels, word mapping, and meta data tags
for data files that will be matched based on column header. Use case for
this is mostly for creating pipelines where it is known what data can be
expected from source data, but matching on filename and column number
is not reliable enough.